+++ android/src/main/kotlin/com/wsc/sim_card_info/SimCardInfoPlugin.kt
@@ -70,27 +70,37 @@ class SimCardInfoPlugin : FlutterPlugin, MethodCallHandler, ActivityAware,
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
             val subscriptionManager = context.getSystemService(SubscriptionManager::class.java)
             subscriptionManager?.activeSubscriptionInfoList?.let { subscriptionInfoList ->
-               
+
                 for (info in subscriptionInfoList) {
                     writer.beginObject()
                     writer.name("carrierName").value(info.carrierName.toString())
                     writer.name("displayName").value(info.displayName.toString())
                     writer.name("slotIndex").value(info.simSlotIndex.toString())
-                    writer.name("number").value(info.number.toString())
+
+                    if (info.number != null || info.number != "") {
+                        writer.name("number").value(info.number.toString())
+                    } else {
+                        writer.name("number").value("")
+                    }
                     writer.name("countryIso").value(info.countryIso.toString())
                     writer.name("countryPhonePrefix").value(info.countryIso.toString())
                     writer.endObject()
-                   
+
                 }
                 writer.endArray()
-                
+
             }
         } else {
              writer.beginObject()
             writer.name("carrierName").value(telephonyManager.networkOperatorName.toString())
             writer.name("displayName").value(telephonyManager.simOperatorName.toString())
             writer.name("slotIndex").value(telephonyManager.simSerialNumber.toString())
-            writer.name("number").value(telephonyManager.line1Number.toString())
+
+            if (telephonyManager.line1Number != null || telephonyManager.line1Number.toString() != "") {
+                writer.name("number").value(telephonyManager.line1Number.toString())
+            } else {
+                writer.name("number").value("")
+            }
             writer.name("countryIso").value(telephonyManager.simCountryIso.toString())
             writer.name("countryPhonePrefix").value(telephonyManager.simCountryIso.toString())
             writer.endObject()