+++ lib/sim_info.dart
@@ -38,6 +38,24 @@ class SimInfo {
     required this.countryPhonePrefix,
   });
 
+  SimInfo copyWith({
+    String? carrierName,
+    String? displayName,
+    String? slotIndex,
+    String? number,
+    String? countryIso,
+    String? countryPhonePrefix,
+  }) {
+    return SimInfo(
+      carrierName: carrierName ?? this.carrierName,
+      displayName: displayName ?? this.displayName,
+      slotIndex: slotIndex ?? this.slotIndex,
+      number: number ?? this.number,
+      countryIso: countryIso ?? this.countryIso,
+      countryPhonePrefix: countryPhonePrefix ?? this.countryPhonePrefix,
+    );
+  }
+
   factory SimInfo.fromJson(Map<String, dynamic> json) {
     return SimInfo(
       carrierName: json['carrierName'],
@@ -54,17 +72,27 @@ class SimInfo {
     return 'SimInfo{carrierName: $carrierName, displayName: $displayName, slotIndex: $slotIndex, number: $number, countryIso: $countryIso, countryPhonePrefix: $countryPhonePrefix}\n';
   }
 
+  Map<String, dynamic> toJson() {
+    return {
+      'carrierName': carrierName,
+      'displayName': displayName,
+      'slotIndex': slotIndex,
+      'number': number,
+      'countryIso': countryIso,
+      'countryPhonePrefix': countryPhonePrefix,
+    };
+  }
+
   @override
   bool operator ==(Object other) =>
-      identical(this, other) ||
       other is SimInfo &&
-          runtimeType == other.runtimeType &&
-          carrierName == other.carrierName &&
-          displayName == other.displayName &&
-          slotIndex == other.slotIndex &&
-          number == other.number &&
-          countryIso == other.countryIso &&
-          countryPhonePrefix == other.countryPhonePrefix;
+      runtimeType == other.runtimeType &&
+      carrierName == other.carrierName &&
+      displayName == other.displayName &&
+      slotIndex == other.slotIndex &&
+      number == other.number &&
+      countryIso == other.countryIso &&
+      countryPhonePrefix == other.countryPhonePrefix;
 
   @override
   int get hashCode =>