+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 0.0.1
+## 1.0.2
 
+* Fix Minor Issues.
+* update Package Description.
+* update dependencies.
+
+## 1.0.1
+
+* Fixed "Invalid json from method channel" issue.
 * TODO: Describe initial release.